+++ src/rust/engine/engine_pyo3/Cargo.toml
@@ -18,14 +18,10 @@ extension-module = ["pyo3/extension-module"]
 default = []
 
 [dependencies]
-either = "1.6"
 fs = { path = "../fs" }
-hashing = { path = "../hashing" }
-itertools = "0.10"
 nailgun = { path = "../nailgun" }
 parking_lot = "0.11"
 pyo3 = "0.15"
-store = { path = "../fs/store" }
 task_executor = { path = "../task_executor" }
 testutil_mock = { package = "mock", path = "../testutil/mock" }
 workunit_store = { path = "../workunit_store" }
